Refactor feature flag initialization
Modularize feature flag initialization by moving it to a new file in the bootstrap package. This decouples the feature flag setup from the main application logic, improving readability and maintainability. Additionally, replace direct NATS calls with a common KV store utility for consistency.
